move `WriteDownPlanViewModel` and `KeyboardController` to separate files
